feat: do not hide resource name flag for update methods
In order to update a resource using CLI we need to specify it's resource
name. AIP does not provide guidance around using OUTPUT_ONLY field
behaviour on a resource's names. There is a use case to have the
OUTPUT_ONLY field behaviour annotation set on resource names in protos
so it makes sense to ignore this in the CLI generator.