feat: add support for passing options + global token function
In order to support overriding how a bearer token is retrieved we add
support for specifying options to module commands with the single option
supported in this PR being a function where the user of this library can
override how a bearer token is retrieved on a global level.

A couple of interesting decisions were taken in this commit which I'll
attempt to document below:

Since a CLI will generally have its entrypoint on a module command, it
made sense to set the options argument on the constructor function for
that and since the existing API `NewModuleCommand` function already has
a variadic argument, in order to avoid a breaking change, we define a
new `Module` struct with a constructor that takes optional arguments and
then define a `Command` method on that which follows the same API as the
previous `NewModuleCommand` function, adjusting the generated code
accordingly.

In order to avoid complex changes, we set the options in the command
context rather than trying to define a shared state struct which would
mean a lot of chnages in the code base. While not the most Go idiomatic
way of doing things it works in this context and seems to be what the
cobra community has done at times to overcome lack of support for such a
setup in Cobra.

We also want options to be able to affect all commands of the final CLI
(and maybe in the future only specific children commands) but the module
command constructor can be passed as a child of other module commands.
This means that if we use the direct context of the command the option
has been set on, it would not be possible to affect a sibling or a
parent command and all their children so ultimately what we do is that,
at runtime, we "merge" all options found in all commands and their
children into the final option which gets put on the root command and
those are the options that are used. How exactly each option is merged
and then used is up to each option to define. This allows us to perform
validation on each option in cases when they have been set more than
once and they are not meant to and also to support cases in the future
where an option only affects its command and/or its children rather than
every command.

Lastly, we introduce a new module command option `WithGlobalTokenFunc`
which allows overriding how a bearer token is retrieved before gRPC
service are called globally for all commands. If not supplied the
current behaviour is kept.