feat: add support for passing options + global token function
In order to support overriding how a bearer token is retrieved we add
support for specifying options to module commands with the single option
supported in this PR being a function where the user of this library can
override how a bearer token is retrieved on a global level.

A couple of interesting decisions were taken in this commit which I'll
attempt to document below:

Since a CLI will generally have its entrypoint on a module command, it
made sense to set the options argument on the constructor function for
that and since the existing API `NewModuleCommand` function already has
a variadic argument, in order to avoid a breaking change, we define a
new `Module` struct with a constructor that takes optional arguments and
then define a `Command` method on that which follows the same API as the
previous `NewModuleCommand` function, adjusting the generated code
accordingly.

In order to avoid complex changes, we set the options in the command
context rather than trying to define a shared state struct which would
mean a lot of chnages in the code base. While not the most Go idiomatic
way of doing things it works in this context and seems to be what the
cobra community has done at times to overcome lack of support for such a
setup in Cobra.

Each command can have its own defined options stored in its own context
but we also want be able to set "global" options, meaning options that
can affect all commands. In order to do that we setup some helper
functions allowing walking down the command chain and find an option
that has been set down that hierarchy.

Lastly, we introduce a new module command option `WithGlobalTokenFunc`
which allows overriding how a bearer token is retrieved before gRPC
service are called globally for all commands. If not supplied the
current behaviour is kept.