feat: add support for optional
Optional fields in the resource bodies are handled by the json
marshalling/unmarshalling, so they don't have to be handled explicitly.

Optional fields in the requests instead are mapped to command line flags
(e.g. `--optional-field-name`), and by default are always mapped to
default values (`""`, `0`, ...) when the flag is not provided.

With this PR, the value of command line flags corresponding to optional
fields is:
- set to `nil` if the flag was missing
- kept to its value (including the default value) if a value is set

PE-1229